requirements.txt
cohere==5.1.8
langchain==0.1.17
langchain-core==0.1.51
langchain-community==0.0.36
langchain-pinecone==0.0.3
langchain-cohere==0.1.2
pinecone-client==3.2.2
flask==3.0.3
python-dotenv==1.0.1
packaging==23.2
pypdf==3.17.1
